chore: update queue type for bulk creation from production plan

+++ erpnext/manufacturing/doctype/production_plan/production_plan.py
@@ -910,7 +910,7 @@ class ProductionPlan(Document):
 
 		frappe.enqueue(
 			self.create_all_work_orders_and_job_cards_for_production_plan,
-			queue="long",
+			queue="short",
 			user=frappe.session.user if frappe.session else None,
 		)
